Refactor, add user link factory
