+++ src/DataFixtures/PositionDictionaryFixtures.php
@@ -5,9 +5,8 @@ declare(strict_types=1);
 namespace App\DataFixtures;
 
 use App\Entity\User;
-use App\Service\Factory\PositionDictionaryFactory;
+use App\Service\Factory\Dictionary\PositionDictionaryFactory;
 use Doctrine\Bundle\FixturesBundle\Fixture;
-use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\DataFixtures\DependentFixtureInterface;
 use Doctrine\Persistence\ObjectManager;
