+++ src/Service/Factory/UserFactory.php
@@ -16,11 +16,6 @@ class UserFactory
         ?string $surname,
         string $password,
         array $roles,
-        string $githubLink = null,
-        string $gitlabLink = null,
-        string $linkedinLink = null,
-        string $websiteLink = null,
-        string $youtubeLink = null,
     ): User {
         $user = new User();
 
@@ -29,11 +24,6 @@ class UserFactory
         $user->setSurname($surname);
         $user->setPassword($passwordHasher->hashPassword($user, $password));
         $user->setRoles($roles);
-        $user->setGithubLink($githubLink);
-        $user->setGitlabLink($gitlabLink);
-        $user->setYoutubeLink($youtubeLink);
-        $user->setWebsiteLink($websiteLink);
-        $user->setlinkedinLink($linkedinLink);
 
         return $user;
     }
